Add init m_buy_points

+++ server_app/strategy_buy_spot/strategy_buy_buy_point.h
@@ -15,7 +15,7 @@ struct BuyPoint
 
     double price = 0.0;
     double quantity = 0.0;
-    size_t profit = 0.0;
+    double profit = 0.0;
     Status status = Status::AVAILABLE;
 
     static std::string to_string(Status data)
